docs: cross kernel snapshot incompatibility on AMD
We recently introduced a check (#3164) to ensure that when restoring a
Firecracker uVM from a snapshot we are able torestores the values of
guest MSRs. If not all MSRs were restored properly Firecracker aborts,
to avoid undefined behaviour.

This has revealed that on AMD a snapshot taken on a 5.10 kernel will not
properly restore on 4.15 kernel, so this commit updates the
compatibility matrix in the docs.

Signed-off-by: Babis Chalios <bchalios@amazon.es>