Revert default behavior of cargo build
My PR #3213 carelessly changed the default behavior of cargo build.
This commit restored the old behavior while still fixing
cross-compilation. For this, instead of relying on the default
target matching the compilation host, we instead explicitly read
the host target from the HOST variable that cargo sets
for build scripts.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>