test: reduce snapshot size in test
`test_5_inc_snapshots` test has been intermittently failing due to
timeouts. The reason is that the snapshot we're taking is quite big and
causes the test to reach its timeout.

However, the size of the VM does not really matter in what we are
testing. This commit reduces the memory size of the VM from 4GB to 512MB
in order to avoid timeouts.

Signed-off-by: Babis Chalios <bchalios@amazon.es>