CI: fix 5_inc_snapshots intermittent failure
We have stumbled upon an OOM scenario in one of our integration
tests (5_inc and 5_full snapshots) due to the fact that we spawn FIO
in the guest in every create/resume snapshot sequence.
Doing it upon each invocation of snap/resume might possibly lead to
multiple FIO instances on the guest (the FIO processes from the
previous guest might get snapshotted). This can result in an OOM and
thus lead to an intermittent failure in the CI.

Signed-off-by: Diana Popa <dpopa@amazon.com>