Add a regression test
The added rust integration test causes the CI to fail on PRs that
introduce new dependencies using comparison requirements.

It is a rust integration test so that we can make use of library
code for robust parsing of Cargo.toml files instead of rolling
a hand-written, brittle toml parser.

it is placed in the src/firecracker crate since it is not possible
to have workspace-wide rust integration tests.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>