api_server: fix error propagating
The old IOError and EventFdError in the api_server's
enum Error are not actually used.
They have been at some point used but we later introduced
the logci where bind_and_run would call std::proces::exit on
errors on HttpServer creation so in reality the IOError and
EventFD error would not be thrown ever.

There are still some things to figure out: bind_and_run will
not return any type of error except for the one we add in this
commit. All the error processing happens inside bind and run
 in reality.

Signed-off-by: Diana Popa <dpopa@amazon.com>