Fix test_dependencies only checking dependency cnt
This allowed dependencies to be replaced without the test failing, as
long as the overall number of lines in the dependencies.txt file was
correct.

Note that due to the use of pprint automatically wrapping lines at 80
characters, the file contained more lines than dependencies, making
the check even more inaccurate (dependencies with long paths/git urls
werevbroken up into multiple lines, leading to a leniency of two
dependencies that could be added without the test failing).

This commit changes to the to
1. fix the above issues
2. ignore dependency versions, and only care about their names
   (meaning updates to a dependency will not trigger the test)

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>