Improve robustness of sanitize_process()
The previous approach was to iterate through the known attached file
descriptors, we now instead iterate through and attempt to close all
possible attached file descriptors. This offers a more robust
and safe implementation.

This fixes a bug which arises in Rust versions >=1.65.0 from a file
descriptor being closed twice.

Signed-off-by: Jonathan Woollett-Light <jcawl@amazon.co.uk>