chore(test): use record_property instead of returning from tests
Returning from a test is deprecated in pytest

```
build/test_binary_size.py::test_firecracker_binary_size
PytestReturnNotNoneWarning: Expected None, but returned
('2505192 B', '2520632 +/- 5.0% B'), which will be an error in a future
version of pytest.  Did you mean to use `assert` instead of `return`?
```

This was used by the test reporting framework to augment test results.
We will switch to a more mainstream way of doing test reports in a
subsequent commit.

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>