chore: Update snap/restore baseline
The snapshot/restore performance test continued to fail since
2022-10-30. As a result of investigation, ioctl() for KVM_CREATE_VM
and the first KVM_CREATE_VCPU (especially sending IPIs to all-but-self
on xAPIC environment) takes much time and we conclude that this latency
came from APIC setting (not from firecracker side) and we cannot handle
it. Then, this commit updates baselines for that.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>