[CI][mem_monitor]: fix bug
The way the memory monitor works right now is
by spawning a separate thread while integration testing
takes place and monitoring Firecracker's memory usage.

When the amount exceeds a certain threshold, we add that
to a queue. This queue is processed when "kill" is called on the
microVM. The condition for checking the contents of the queue
is for the memory monitor to have been initialized and for it
to be alive.
The latter requirement is incorrect since when the exceed happens,
we exit the "run" thread associated function. The exit of the "run"
will trigger the thread to enter the "stopped" state (i.e not "alive").

Signed-off-by: Diana Popa <dpopa@amazon.com>