test(virtio): add a test helper for virtio devices
Add a test struct to allow testing virtio devices. The struct emulates
the functionality of the guest driver, along with the event manager on
the Firecracker side.

We actually use the same approach for the Network device. This commit
abstracts away the device specific implementation details and provides a
common API for all devices to use.

We will use this initially for the virtio-rng device which we are
introducing and we will later start using it for the rest of the devices
as well.

Signed-off-by: Babis Chalios <bchalios@amazon.es>