feat(virtio): add an IoVecBufferMut type
We already have an IoVecBuffer type which allows us to use `writev` in
the TX path of net devices. Apart from allowing us to do scatter-gather
operations, this type gives us a clean interface to parse a descriptor
chain into a slice of slices.

So far, we have only the immutable variant of this type. This commit
adds a mutable variant (`IoVecBufferMut`), so we can create a view of a
write-only descriptor chain.

Also, move the type definition out of the `net` device module and into
the top-level `virtio` module so it can be used by other devices as
well.

Signed-off-by: Babis Chalios <bchalios@amazon.es>