feat(virtio): add iterator type for DescriptorChain
In virtio code its very often that we want to iterate over all the
descriptors that are part of a chain and describe a buffer that the
guest passed to us. To do so, we make use of the `DescriptorChain`
methods `has_next()` and `next_descriptor()`.

This commits adds an `Iterator` type that allows us to iterate over
these descriptors in a more Rust-idiomatic way.

Signed-off-by: Babis Chalios <bchalios@amazon.es>