fix(virtio): undefined behaviour in net `writev` implementation
The way we implemented `IoVecBuffer` was leading to potentially
undefined behaviour. We were creating a normal slice from guest physical
memory. Although, with well-behaved drivers this should be fine, a
driver that touches that memory after adding the buffer in the queue and
before the device adds it in the used queue, would cause undefined
behaviour due to the aliasing requirements of Rust for slices.

This commit avoids making a slice out of raw guest physical memory for
`IoVecBuffer`s. It instead uses directly `libc::iovec` to describe the
buffers. As a consequence, in order to have vectored writes for tap
devices we need to call directly `libc::writev`.

Signed-off-by: Babis Chalios <bchalios@amazon.es>