tests: override record_property
For tests that set different dimension values within the same test (like
some of the performance tests do), all the metrics published except for
the last one would be missing the properties, since that happens during
fixture teardown.

We can simply move it before, for the properties we know, but any
properties recorded by later fixtures could be lost. Fortunately pytest
allows overriding a fixture[^1]. now every time we record a property, we
also set a property in the metrics logger.

[^1]: https://docs.pytest.org/en/7.1.x/how-to/fixtures.html#override-a-fixture-on-a-folder-conftest-level

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>