feature: uffd: implement configurable prefault
Prefaulting can be customized through `--pages-after` optional
argument. The prefaulter will copy into memory the range:
[faulting_page_start, faulting_page_end + bytes_after],
where:
bytes_after = pages-after * page_size

If the range is not within the current region, the chunk is
truncated to fit into the current region.

Note that there is no guarantee that the handler will always
fetch the requested number of pages.  Starting from the
faulting page, the handler will copy pages until it finds one
that has already been brought into memory (during a previous
page fault) or until a number of `pages-after` pages have been
copied.

For starters, the start address to bring into memory is the
faulting page.

End address to prefault is computed from the end of the faulting
page as follows:
- Add `after_kib` bytes to the end of the faulting page. If it
  overflows, default value is end of the current region.
- Make sure that the address fits inside the current region.

Signed-off-by: Luminita Voicu <lumivo@amazon.com>