refactor: uffd: move GuestRegionUffdMapp to utils
When userspace is in charge of handling page faults, Firecracker
sends its guest memory mapping to the userfaulfd handler process.
The mappings are represented through `GuestRegionUffdMapping`
structure, which is defined in both vmm and uffd handler logic.
This commit moves `GuestRegionUffdMapping` to the utils crate,
in order to remove duplicate code and be able to use the same
structure in both crates. This also has the advantage that any
change to GuestRegionUffdMapping will be reflected in both
Firecracker and uffd handler.

Signed-off-by: Luminita Voicu <lumivo@amazon.com>