feature: uffd: kill Firecracker if uffd crashes
Firecracker relies on UFFD handler to serve page faults. In case
the UFFD handler encounters an error and crashes, Firecracker will
freeze waiting to have an userspace process copy the requested
page into RAM. There is no way to signal to kernel that it needs
to take responsibility to handle faults. What we can do to prevent
Firecracker from stalling is to send it a SIGBUS signal from uffd
handler, just before it crashes.

Signed-off-by: Luminita Voicu <lumivo@amazon.com>