test: check that snapshots can be taken at current version
Add a test that verified that taking a snapshot with target_version set
to the version specified in src/firecracker/Cargo.toml works, and yields
a snapshot where --describe-snapshot returns that very same version
again.

This test covers the short window between a version bump before a
release during which no new release binary has been uploaded to S3 yet.
Previously, the existing tests would not catch if it was not possible to
take a snapshot in the about-to-be-released version, as only versions
that had their binaries in S3 were considered.

Co-Authored-By: Pablo Barbáchano <pablob@amazon.com>
Signed-off-by: Patrick Roy <roypat@amazon.co.uk>