test: Remove  `test_ensure_mod_tests`
The current implementation of the test does not cover most error cases.

A full solution would need to consider the Rust AST. In the worst case
a test gets included in coverage and raises the coverage slightly
higher than it should be, this is not a significant impact.

The test does check some cases, this may be better than no test at all,
however given the impact it is not significant enough to warrant the
complication.

Signed-off-by: Jonathan Woollett-Light <jcawl@amazon.co.uk>