gitlint: Ignore dependabot commits
Dependabot commit messages are constructed by including the changelog of
the package for which dependabot PRs in an update. We have no control
over how these packages write their commits messages, so just disable
gitlint for them.

Dependabot does not seem to have an option for configuring the commit
messages beyond the title line. However, even if there was such an
option, the next problem would be that GitHub's security update PRs
(which are separate from normal dependabot) ignore the
`.github/dependabot.yml` configuration file.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>