test(entropy): add functional integration tests
When a microVM has been configured with a virtio-rng device the
character device `/dev/hwrng` uses this as a source of random numbers.
Add a negative control test that ensures we can't get any random bytes
out of this device if virtio-rng has not been configured and one that
ensures that we can when virtio-rng is there.

These tests are skipped for c7g instances where we need to have a no SVE
kernel for host kernels 4.14. This is because we are using for these
tests a `test_microvm_with_rng` fixture which is not flexible as to what
kind of kernel we can use per instance type.

Signed-off-by: Babis Chalios <bchalios@amazon.es>