fix(vmm): undo removal of 0xb leaf on AMD
The removal of 0xb leaf on AMD was introduced in PR #2164, derived from
issue #222. As seen in the issue, the reporter told that upgrading guest
kernel from 4.19.129 to 4.19.146 fixed the issue, meaning that this was
not a KVM (host kernel) side issue.

Actually, leaf 0xb is valid even on AMD, as described in "E.3.8 Function
Bh — Extended Topology Enumeration" section of AMD64 Architecture
Programmer's Manual Volume 3. And the meaning of each bit is consistent
with Intel.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>