tests: improve on dependency test
The test for checking the number of dependencies
stays the same (`test_num_depdendencies`) would only check
whether the number of the current ones is not higher than
the previous dependencies.
We would lack the logic for making sure every time we remove a
dependency the file keeping track of the dependencies is also updated.

Signed-off-by: Diana Popa <dpopa@amazon.com>