refactor(vmm): move MSR modification for linux boot
Move MSR modification to boot linux from `CpuConfiguration::host()` to
`KvmVcpu::configure()`.

As CPUID modification is applied in `KvmVcpu::configure()` already,
putting MSR modification in the same place is sensible. In addition,
with this change, there is no need to contain `msrs_to_save` and
`msr_boot_entries` in `CpuConfiguration`.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>