refactor(vmm): 2-step CpuConfiguration build for x86_64
Build a custom `CpuConfiguration` with 2 steps:
1. Construct a base `CpuConfiguration`.
2. Apply CPU template to the default `CpuConfiguration`.

Previously, all these processes were done in `CpuConfiguration::new()`,
but this approach is a bit complex in the test perspective. In addition,
we will be able to unify `CpuConfiguration` interface between x86_64 and
aarch64 and simplify `configure_system_for_boot()` function in
successive commits.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>