feat(vmm): add vendor check of CPU template
Check whether the vendor is matched between the static CPU template
and the actual CPU.

Static CPU templates for x86-64 are specific to vendor (Intel or AMD).
`include_leaves_from()` method does this check now, but will be removed
in a successive commit. Instead of this method, we do this check in
`get_cpu_template()`.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>