feat(tool): add CPU template helper tool
Add a helper tool that make it easy to create and maintain custom CPU
templates.

Custom CPU template needs to be crafted carefully, because they give
customers a great ability to modify CPU features exposed to guests that
can result in guest crashes or undefined behaviors. As firecracker
relies on KVM largely and does not have runtime checks for given custom
CPU templates, customers have the responsiblity of their custom CPU
templates. To allow customers to create and maintain as easily as
possible, a helper tool is added as a separate binary from firecracker.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>