test: stop CPU load monitoring at the end of a test
We see errors during teardown. It is then not clear if the CPU usage
happened during the call phase of the test (which would be legit) or the
teardown (which would be a false positive).

Restrict the CPU load monitoring part to only the part of the test we
are interested in.

Also remove the CPU load monitoring from Microvm, since this is the only
user and it fits better as a Context Manager.

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>