feat(tool): Add remove_common() for strip feature
Add `remove_common()` to retrieve common items shared by all the given
list of `HashSet` and remove the shared items from them.

Implementation for converting a list of CPU templates
`Vec<CustomCpuTemplate>` to a list of sets of modifiers
`Vec<HashSet<Modifier>>` is required for each architecture (x86_64 and
aarch64), because they have different structures for
`CustomCpuTemplate`. But the core part to remove shared modifiers can be
the same logic and this commit adds it.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>