feat(tool): Apply register-level strip
Strip shared CPUID modifiers with a granularity of register.

Previously the strip feature for CPUID modifiers work in leaf-wise,
because the template format for CPUID has a nested structure where leaf
modifiers have register modifiers inside and it is not straightforward.
This commit introduces a new data structure `CpuidModifier` that has a
flatten modifier and implements `From` trait for type conversion.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>