refactor(tool): Make dump() return CustomCpuTemplate
Make `dump()` return `CustomCpuTemplate` instead of serialized `String`
by moving JSON serialization to `main.rs`.

The dump functionality will be re-used in a feature to verify that the
given custom CPU templates are applied as intended. By moving the
serialization out, the verify feature does not need to deserialize it
again to run verification for it.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>