refactor(vmm): remove Leaf0NotFoundInCpuid error
Leaf0NotFoundInCpuid was only used by validate_cpu_vendor()
as an inner error type.
Removing it as it does not add any extra information
on top of the outer ValidateCpuVendorError::Snapshot error.
Moreover, validate_cpu_vendor() is not even the right
level to tell if the error was caused by absence of the leaf 0.

Signed-off-by: Nikita Kalyazin <kalyazin@amazon.com>