test: add GitHub Action that verifies no lockfiles are dirty
The `cargo metadata` command would generally do the same changes to a
lockfile as a `cargo check` (which we cannot run in a GitHub action as
it runs compiler passes on our dependencies). Passing `--locked` means
that instead of updating a lockfile, it will print an error and exit
with non-zero code.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>