chore: snapshot latency baselines for x86_64
We have been observing performance failures on the snapshot restore
latency tests mainly on `m5d` but also a few cases on `m6i` and `m6a`.

Potential root causes are AMI updates since the last time we gathered
baselines as well, the Rust toolchain update and a security fix on the
`versionize` crate.

Signed-off-by: Babis Chalios <bchalios@amazon.es>