Use Read/WriteVolatile for block device sync io
The vm-memory provided read/write_at methods taking a generic Read/Write
impl and using that to transfer data from/to guest memory have recent
introduced a copy-to-vmm-memory as part of the transfer, due to
soundness concerns of using rust-slices to operate on guest memory. The
Read/WriteVolatile traits avoid this copy by operating directly on
`AsRawRd` implementations.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>