refactor: `guest_config` into `cpu_config`
This refactoring is done to reduce nesting of modules inside
`guest_config` crate and give it a more proper name.

Reason for renaming:
Originally `guest_config` was intended as a module with separate modules
inside, each corresponding to separate device that can be configured
by the user. But as of right now it only contains CPU related configs
(CpuConfiguration, Static/CustomCpuTemplates, etc...).

Reason for restructuring:
Because Firecracker supports x86_64 and aarch64 `guest_config`
is divided into several `x86_64` and `aarch64` modules at
different depth levels.

This is how it looks now:
  src/vmm/src/guest_config/
  ├── aarch64
  │   ├── mod.rs
  │   └── static_cpu_templates
  ├── mod.rs
  ├── templates
  │   ├── aarch64.rs
  │   ├── mod.rs
  │   ├── test_utils
  │   │   ├── aarch64.rs
  │   │   ├── mod.rs
  │   │   └── x86_64.rs
  │   └── x86_64.rs
  ├── templates_serde.rs
  └── x86_64
      ├── cpuid
      ├── mod.rs
      └── static_cpu_templates

There are 3 levels of `x86_64`/`aarch64` separation.

With this refactoring tree looks like this:
  src/vmm/src/cpu_config/
  ├── aarch64
  │   ├── custom_cpu_template.rs
  │   ├── mod.rs
  │   ├── static_cpu_templates
  │   └── test_utils.rs
  ├── mod.rs
  ├── templates.rs
  ├── templates_serde.rs
  └── x86_64
      ├── cpuid
      ├── custom_cpu_template.rs
      ├── mod.rs
      ├── static_cpu_templates
      └── test_utils.rs

Only 1 level of `x86_64`/`aarch64` separation.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>