feat(entropy): use aws-lc-rs for random bytes
aws-lc-rs is a cryptographic library based on the aws-lc native library.
Currently, we are using OsRng from rand crate for getting random bytes
to fulfill guest requests.

This commit switches from rand crate to aws-lc-rs. The latter is in the
process of getting FIPS certification which might be of use to use in
the future.

Signed-off-by: Babis Chalios <bchalios@amazon.es>