docs: refer to linux docs in prod host setup guide
Mitigations against side channel attacks listed in the guide
are not exhaustive and there always is a room to add more.
Instead of attempting to keep up with state-of-the-art
guidelines, the majority of mitigations are replaced with
links to Linux kernel and CPU vendor documentation.
Also, it is emphasised that responsibility for those
mitigations lies on the host configuration and out of
Firecracker control.

Signed-off-by: Nikita Kalyazin <kalyazin@amazon.com>