fix integer overflow/excessive token regeneration in auto_replenish
Due to rounding down instead of up when computing the time that it would
take to generate "x" tokens, the fractional part of a nanosecond is used
"twice" once per call to auto_replenish. This can only happen if the
time since the last auto_replenish call is enough to generate at least
one token, e.g. (refill_time / bucket_size) milliseconds. To accumulate
enough "double used" time to generate an extra token, at least
(refill_time * 1_000_000 / bucket_size) calls to auto_replenish are
needed, with the calls timed in a way that causes the "double used" time
to be as close as possible to 1ns. Note that since time resolution is at
the nanosecond level, at least one nanosecond has to pass between these
calls.

For example, if a TokenBucket is setup for 1GB/s, we could allow one
extra byte of traffic every (1000 / 1_000_000_000) * 1_000_000 = 1ms,
e.g. excess throughput of 1KB/s (or 0.0001% above the target)

Additionally, in extremely unrealistic scenarios (either refill_time of
thousands of years or the guest waiting thousands of years), an integer
overflow in a multiplication in auto_replenish can leads to less tokens
being replenished than should be.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>