fix(vmm): Disable perfmon and debug capability
We completely disabled the PMU on CPUID.0AH in 58b71f073cea ("cpuid:
disable performance monitor unit") to prevent an attacker from
redirecting control flow to malicious code.

On the other hand, Firecracker left CPUID:01H.ECX[bit 15] as it is,
which indicates the processor supports the performance and debug feature
indication IA32_PERF_CAPABILITIES MSR enumerating the existence of
performance monitoring features.

Thit bit is set to 1 on Intel CPUs + kernel 5.10 by default. And now
the custom CPU template feature allows users to set to 1. To prevent
users and newer kernels from enabling it, this commit disables it in
CPUID normalization process.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>