test: do not track iops in block perf test
Analysis of historical performance test data has shown that iops and
bandwith are correlated as follows:

bandwidth = iops * block_size

with a pearson correlation coefficient of 1. This makes sense, as we
collect both bandwidth and iops from fio. Here, each I/O operation is
either a read or a write of a fixed size.

This means explicitly tracking both does not make sense.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>