test: remove read and readwrite fio modes from block perf test
Analysis of historical data has shown that the read and randread and
readwrite and randrw have highly correlated results (r=0.96 and r=0.97,
respectively). Therefore, remove the sequential ones.

In a later commit, we'll replace the "randrw" with "randwrite" (as fio
tries to perfectly balance read and write throughput for the
readwrite/rw tests, meaning we really only measure read OR write by
using the combined mode (whatever happens to have the lower
throughput)).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>