test: Only use default window size for tcp throughput perf test
Analysis of our baselines shows that in all cases, the highest
throughput is achieved using the default window size. As there is no
code in firecracker that gives us reason to believe we'd need to
specifically monitor other windows sizes, and our performance tests are
intended to profile firecracker's theoretical maximum performance, there
is no need to track these sub-optimal configurations.

A later PR could fine-tune the window size (and other paramters) to
ensure we actually use the best possible value for it, but that would
require baseline regathering and it out of scope for now.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>