refactor(vmm): Remove aarch64 core register exception
Since the kernel 4.14.312 or older has a bug on KVM_GET_ONE_REG, we
added a temporary mitigation to dump the right aarch64 core registers in
PR #3586.

Now that our CI is using 4.14.314 and there is enough time for users to
update their kernel version until firecracker 1.5 release, this commit
removes the temporary mitigation.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>