chore: set dependabot to only give us PRs on mondays
Due to the volume of PRs received, and the time wasted due to
rebasing/waiting for CI, we decided to batch the dependabot PRs on
monday, merge each of them onto a temporary branch, and the only merge
that branch to main.

Also increases the PR limit so that we definitely get ALL dependabot PRs
on monday (and dont start falling behind because it only gives them to
us 5 at a time).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>