test: add 'no-block-pr' pytest marker
The marker works identically to the 'nonci' marker, but with different
meaning: It is supposed to be used on tests that will run on every PR,
but whose failure should not block merging.

Additionally, switch from the --nonci option to using pytest's
built-in, marker-based test filtering. This has the advantage that
instead of skipping tests, we deselect them, reducing the log spam due
to skipped tests (which was particularly noticable in the performance
tests).

pytest.ini defines a default argument of `-m 'not nonci and not
no-block-pr'`, which deselects all nonci and optional tests. This can be
overwritten when invoking pytest (through devtool test) by setting -m,
for example to '-m optional' to only run optional tests, or to '-m
nonci' to only run nonci tests.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>