feat(aarch64): ability to work with wide registers
- Updated `kvm-bindings` to 1.4 which changes `get/set_one_reg` methods.
- Updated internal representation of aarch64 registers

Now we are able to work with register wider then 128 bits.
All registers are stored in the `Aarch64RegisterVec` in one byte array.
This way registers of different sizes can be efficiently stored
and iterated on.
Supported sizes are: 8, 16, 32, 64, 128, 256, 512, 1024, 2048 bits
(all sizes supported by KVM).
Register size is determined by the ID of the register.
Interactions with registers are happening using `Aarch64RegisterRef`
or `Aarch64RegisterRefMut`.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>